Working to refactor project into client/server. Adding Protobufs
