+++ client/printer/printer.go
